Update TextraTypist to latest commit; prepare for next release. (#131)
* Use makeTooltip() instead of fixTooltip().

I still have fixTooltip here, but there may turn out to be more involved fixes necessary at creation time. If we encounter those, they can be fixed in one place.

* Add emoji to all Fonts in a family.

This makes it so the emoji selector works regardless of which Font in the family is currently active. Before, it only worked for the first Font in the family.

* Change TT dep, some fonts along with.

* Stripe's color picker includes an old TT.

I use `api` here, but this could be swapped back to `compile` and the plain `java` plugin without issue (it's just that api will actually work on future, maybe current Gradle, while compile at some point won't). I added an exclude on the TextraTypist the color picker uses, so I hope things are fully compatible in 0.7.5 with its expected 0.6.3 .