Update emoji-related files.
There are some new emoji here, like jellyfish and broken chains, as well as some significant improvements to organization, like the bomb emoji being in tools rather than emotion. Twemoji is now the same version as in the latest commit of TextraTypist, but if you use this .atlas and .png, it should work fine in the latest stable or older versions, too.